feat: add indexed scans and database workbench

+++ minidb-engine/src/main/java/minidb/storage/index/IndexEntry.java
@@ -1,7 +1,7 @@
 package minidb.storage.index;
 
 /** One key-to-row mapping returned by an index range scan. */
-public record IndexEntry(int key, RowId rowId) {
+public record IndexEntry(long key, RowId rowId) {
     public IndexEntry {
         if (rowId == null) {
             throw new NullPointerException("rowId");
